LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe-allowfullscreen-expected.txt

CONSOLE MESSAGE: Permission policy 'Fullscreen' check failed for document with origin ''.
CONSOLE MESSAGE: Permission policy 'Fullscreen' check failed for document with origin ''.
CONSOLE MESSAGE: Permission policy 'Fullscreen' check failed for document with origin ''.

PASS iframe-same-origin-allowfullscreen
PASS iframe-cross-origin-allowfullscreen
PASS iframe-noload-noallowfullscreen
PASS iframe-noload-allowfullscreen
